refactor sync manager and add shutdown channel
